Add a CountDifferentPixels function so that we can compare non-numeric images to each other.

+++ ITKHelpers.cpp
@@ -605,7 +605,8 @@ std::vector<itk::Index<2> > Get4NeighborIndicesInsideRegion(const itk::Index<2>&
   return indices;
 }
 
-itk::ImageRegion<2> GetInternalRegion(const itk::ImageRegion<2>& wholeRegion, const unsigned int patchRadius)
+itk::ImageRegion<2> GetInternalRegion(const itk::ImageRegion<2>& wholeRegion,
+                                      const unsigned int patchRadius)
 {
   unsigned int width = wholeRegion.GetSize()[0];
   unsigned int height = wholeRegion.GetSize()[1];
@@ -618,10 +619,11 @@ itk::ImageRegion<2> GetInternalRegion(const itk::ImageRegion<2>& wholeRegion, co
   return region;
 }
 
-std::vector<itk::ImageRegion<2> > GetPatchesCenteredAtIndices(const std::vector<itk::Index<2> >& indices, const unsigned int patchRadius)
+std::vector<itk::ImageRegion<2> > GetPatchesCenteredAtIndices(const std::vector<itk::Index<2> >& indices,
+                                                              const unsigned int patchRadius)
 {
   std::vector<itk::ImageRegion<2> > regions(indices.size());
-  for(int i = 0; i < indices.size(); ++i)
+  for(size_t i = 0; i < indices.size(); ++i)
   {
     regions[i] = GetRegionInRadiusAroundPixel(indices[i], patchRadius);
   }
@@ -645,7 +647,8 @@ std::vector<itk::ImageRegion<2> > GetValidPatchesCenteredAtIndices(const std::ve
   return regions;
 }
 
-std::vector<itk::ImageRegion<2> > GetAllPatches(const itk::ImageRegion<2>& fullImageRegion, const unsigned int patchRadius)
+std::vector<itk::ImageRegion<2> > GetAllPatches(const itk::ImageRegion<2>& fullImageRegion,
+                                                const unsigned int patchRadius)
 {
   typedef itk::Image<float,2> DummyImageType;
   DummyImageType::Pointer dummyImage = DummyImageType::New();
